requirements.txt
python-telegram-bot
motor
dnspython
